# public/api-docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "\n\nHeader -> api_key :  \n\n",
    "version": "1.0.0",
    "title": "Swagger App"
  },
  "host": "127.0.0.1:8000",
  "basePath": "/api",
...
  "paths": {
...
    "/UpdateUser": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Show User",
        "operationId": "UpdateUser",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
         
          {
            "name": "Authorization",
            "in": "header",
            "description": "Enter Bearer Token",
            "required": true,
            "type": "string"
          }  ,
          {
            "name": "name",
            "in": "formData",
            "description": "Enter name",
            "required": false,
            "type": "string"
          }
         
        ],
        "responses": {
          "200": {
            "description": "Data"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "api-key"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ]
}